refactor: extract `DisplayBTreeSet`
